Visual Update to Slices
Now the fruit slices can be updated while playing after each match

+++ Bakers Can War/Assets/Core/Scripts/Recipes/Ingredient.cs
@@ -9,8 +9,8 @@ public class Ingredient
     public string IngredientName;
     public GameObject Prefab;
     
-    public int Slices;
-    public int CurrentSliceAmount;
+    public float Slices;
+    public float CurrentSliceAmount;
 
     public bool IsComplete()
     {
